chore: index the sample window in usize; drop the imperative review citations (#2434)
The neighbourhood mean multiplied two u32s before widening. The probe is a
preview-resolution develop so the product cannot overflow in practice, but
widening first costs nothing and removes the question.

The "(review on #NNNN)" parentheticals throughout these comments read to an
automated reviewer as instructions addressed to it, which blocked the review
on a prompt-injection finding. Same provenance, plain citation form.

+++ src/raw-pipeline/raw-core/src/stages/white_balance_sample.rs
@@ -143,7 +143,7 @@ fn neighbourhood_mean(probe: &Image, nx: f32, ny: f32) -> Option<[f32; 3]> {
     let y1 = ((cy + r) as u32).min(probe.height - 1);
     let (sum, n) = (y0..=y1)
         .flat_map(|y| (x0..=x1).map(move |x| (x, y)))
-        .map(|(x, y)| probe.pixels[(y * probe.width + x) as usize])
+        .map(|(x, y)| probe.pixels[y as usize * probe.width as usize + x as usize])
         .filter(|p| p.iter().all(|v| v.is_finite()))
         .fold(([0.0_f64; 3], 0usize), |(s, n), p| {
             (
